__del__ was not getting called reliably on exit.  As a result, the subprocesses were getting left open (consuming the port).
Everything that I just read about subprocess management, etc, says not to hope for GC to do the work.  most posts think it's better to register the cleanup using atexit().  They also say it's important to not register the cleanup as a system method, which would artificially extend the lifetime of that object.

I've implemented what I perceive to be the recommended version, for your consideration.  This obviously defers the cleanup until the program ends (instead of the object's lifetime ending), but it's more robust.