zmqserver can cache objects
If the zmqserver gets a set_object command that sets precisely the same object data on precisely the same path that already exists, then do not forward it to the connected websockets.
This makes a *dramatic* improvement in the workflow of using meshcat on colab (at least as we use it in drake), because we don't have to repeatedly download large meshfiles from repeated simulations.