Fix meshcat-server start for Python 3.8 on Windows
ZMQ release notes suggested this as the fix for running Python 3.8
on windows, it seems to be a known issue.

I added the run loop policy here as that it is the first thing that
is done inside the subprocess when started.

meshcat-server can now start!

However, running the examples/box.py fails. It seems it cannot
start its own server process. However, if box.py is modified
to use the zmq_url everything works as expected.

A little more work needed here.